requirements.txt
aprs3==8.0.0.post1
attrs==25.3.0
ax253==0.1.5.post1
bitarray==3.3.1
cbor2==5.6.5
importlib_metadata==8.6.1
kiss3==8.0.0
luma.core==2.4.2
luma.oled==3.14.0
pillow==11.1.0
pyftdi==0.56.0
pyserial==3.5
pyserial-asyncio==0.6
pyusb==1.3.1
RPi.GPIO==0.7.1
smbus2==0.5.0
spidev==3.6
zipp==3.21.0
